+++ examples/cs-approval-rules/MyStack.cs
@@ -49,9 +49,6 @@ class MyStack : Pulumi.Stack
                         new EligibleApproverArgs {
                             User = "pulumi-bot"
                         },
-                        new EligibleApproverArgs {
-                            User = "IaroslavTitov"
-                        }
                     }
                 }
             }